// sourcecode/BTrak.Packages/snova-recruitment-module/Recruitment-module-components/recruitment-components/src/lib/snovasys-recruitment-management-apps/models/InterviewProcessConfigurationModel.ts
export class InterviewProcessConfigurationModel {
  isArchived: any;
  timeStamp: any;
  scheduleTimeStamp: any;
  interviewProcessConfigurationId: string;
  interviewTypeId: string;
  interviewTypeName: string;
  jobOpeningId: string;
  candidateId: string;
  statusId: string;
  statusName: string;
  isInitial: boolean;
  interviewProcessId: string;
  roleid: string;
  interviewProcessName: string;
  isPhoneCalling: boolean;
  isVideoCalling: boolean;
  order: number;
  color: string;
  interviewProcessConfigurationIds: any[];
  jobopeningId: any;
}